requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# ML/AI - Transformers and embeddings
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# LoRA fine-tuning support
peft>=0.7.0

# Quantization support (optional, for large models)
bitsandbytes>=0.41.0
accelerate>=0.25.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
lxml>=4.9.0
html2text>=2020.1.16

# LangChain (optional, for advanced RAG)
langchain>=0.1.0
langchain-community>=0.0.10

# OpenAI (optional, for comparison)
openai>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Evaluation metrics
rouge-score>=0.1.2
nltk>=3.8.0

